# Conflicts:
#	src/Nethermind/Nethermind.Api/IApiWithNetwork.cs
#	src/Nethermind/Nethermind.Api/NethermindApi.cs
#	src/Nethermind/Nethermind.Core.Test/Modules/TestMergeModule.cs
#	src/Nethermind/Nethermind.Merge.Plugin/MergePlugin.cs
#	src/Nethermind/Nethermind.Merge.Plugin/Synchronization/BeaconSync.cs
#	src/Nethermind/Nethermind.Optimism/OptimismPlugin.cs
#	src/Nethermind/Nethermind.Synchronization/IBeaconSyncStrategy.cs
#	src/Nethermind/Nethermind.Taiko/TaikoPlugin.cs

+++ src/Nethermind/Nethermind.Benchmark/Store/PatriciaTreeBenchmarks.cs
@@ -351,7 +351,7 @@ namespace Nethermind.Benchmarks.Store
                 Persist.EveryNBlock(2), NullLogManager.Instance);
             StateTree tempTree = new(trieStore, NullLogManager.Instance);
 
-            for (int i = 0; i < _largerEntryCount; i++)
+            for (ulong i = 0; i < _largerEntryCount; i++)
             {
                 if (i % _repeatedlyFactor == 0)
                 {
@@ -359,7 +359,7 @@ namespace Nethermind.Benchmarks.Store
                     tempTree.Commit();
                 }
 
-                (bool isWrite, Hash256 address, Account value) = _largerEntriesAccess[i];
+                (bool isWrite, Hash256 address, Account value) = _largerEntriesAccess[(int)i];
                 if (isWrite)
                 {
                     tempTree.Set(address, value);
