+++ src/Nethermind/Nethermind.Core/ProgressLogger.cs
@@ -20,7 +20,7 @@ namespace Nethermind.Core
         private readonly ILogger _logger;
         private readonly Action<string>? _logAction;
         private string _prefix;
-        private (long, long, long, long) _lastReportState = (0, 0, 0, 0);
+        private (ulong, ulong, long, long) _lastReportState = (0UL, 0UL, 0L, 0L);
         private Func<ProgressLogger, string>? _formatter;
 
         public ProgressLogger(string prefix, ILogManager logManager, ITimestamper? timestamper = null, LogLevel logLevel = LogLevel.Info)
@@ -41,7 +41,7 @@ namespace Nethermind.Core
             };
         }
 
-        public void Update(long value)
+        public void Update(ulong value)
         {
             UtcEndTime = null;
 
@@ -82,7 +82,7 @@ namespace Nethermind.Core
             }
         }
 
-        public void Reset(long startValue, long total)
+        public void Reset(ulong startValue, ulong total)
         {
             LastMeasurement = UtcEndTime = _timestamper.UtcNow;
             UtcStartTime = _timestamper.UtcNow;
@@ -94,7 +94,7 @@ namespace Nethermind.Core
 
         private long _skipped = -1;
 
-        private long StartValue { get; set; }
+        private ulong StartValue { get; set; }
 
         private DateTime? UtcStartTime { get; set; }
 
@@ -102,10 +102,12 @@ namespace Nethermind.Core
 
         private DateTime? LastMeasurement { get; set; }
 
-        private long LastValue { get; set; }
-        public long TargetValue { get; set; }
+        private ulong LastValue { get; set; }
+        public ulong TargetValue { get; set; }
 
-        public long CurrentValue { get; private set; }
+        public ulong CurrentValue { get; private set; }
+
+        // -1 sentinel = queue tracking disabled.
         public long CurrentQueued { get; set; } = -1;
 
         private TimeSpan Elapsed => (UtcEndTime ?? _timestamper.UtcNow) - (UtcStartTime ?? DateTime.MinValue);
@@ -121,7 +123,7 @@ namespace Nethermind.Core
                     return 0M;
                 }
 
-                return (CurrentValue - StartValue) / timePassed;
+                return ((decimal)CurrentValue - StartValue) / timePassed;
             }
         }
 
@@ -160,7 +162,7 @@ namespace Nethermind.Core
                     return 0M;
                 }
 
-                return (CurrentValue - LastValue) / timePassed;
+                return ((decimal)CurrentValue - LastValue) / timePassed;
             }
         }
 
@@ -168,7 +170,7 @@ namespace Nethermind.Core
 
         public void LogProgress()
         {
-            (long, long, long, long) reportState = (CurrentValue, TargetValue, CurrentQueued, _skipped);
+            (ulong, ulong, long, long) reportState = (CurrentValue, TargetValue, CurrentQueued, _skipped);
             if (reportState != _lastReportState)
             {
                 _lastReportState = reportState;
@@ -179,9 +181,9 @@ namespace Nethermind.Core
 
         private string DefaultFormatter() => GenerateReport(_prefix, CurrentValue, TargetValue, CurrentQueued, CurrentPerSecond, SkippedPerSecond);
 
-        private static string GenerateReport(string prefix, long current, long total, long queue, decimal speed, decimal skippedPerSecond)
+        private static string GenerateReport(string prefix, ulong current, ulong total, long queue, decimal speed, decimal skippedPerSecond)
         {
-            float percentage = Math.Clamp(current / (float)(Math.Max(total, 1)), 0, 1);
+            float percentage = Math.Clamp(current / (float)Math.Max(total, 1UL), 0, 1);
             string queuedStr = (queue >= 0 ? $" queue {queue,QueuePaddingLength:N0} | " : " ");
             string skippedStr = (skippedPerSecond >= 0 ? $"skipped {skippedPerSecond,SkippedPaddingLength:N0} Blk/s | " : "");
             string speedStr = $"current {speed,SpeedPaddingLength:N0} Blk/s";
