+++ src/Nethermind/Nethermind.JsonRpc.Test/Modules/Subscribe/TransactionReceiptsSubscriptionTests.cs
@@ -139,7 +139,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_on_receipts_inserted_no_filter_returns_all_receipts()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             Transaction tx1 = Build.A.Transaction.WithHash(TestItem.KeccakA).TestObject;
@@ -170,7 +170,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_on_receipts_inserted_single_hash_filter()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             TransactionHashesFilter filter = new()
@@ -196,7 +196,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_on_receipts_inserted_multiple_hashes_filter()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             TransactionHashesFilter filter = new()
@@ -228,7 +228,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_on_receipts_inserted_non_matching_hashes()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             TransactionHashesFilter filter = new()
@@ -250,7 +250,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_on_receipts_inserted_partial_match()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             TransactionHashesFilter filter = new()
@@ -277,7 +277,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_receipt_includes_all_fields()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).WithHash(TestItem.KeccakF).WithTimestamp(1000000).TestObject;
 
             LogEntry logEntry = Build.A.LogEntry.WithAddress(TestItem.AddressA).WithTopics(TestItem.KeccakA).TestObject;
@@ -319,7 +319,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_logs_have_correct_indices()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             LogEntry log1 = Build.A.LogEntry.WithAddress(TestItem.AddressA).TestObject;
@@ -352,7 +352,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_empty_block_no_notification()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             TxReceipt[] receipts = [];
@@ -366,7 +366,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_failed_tx_still_delivered()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             TxReceipt receipt = Build.A.Receipt
@@ -391,7 +391,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_reorg_skipped()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             TxReceipt receipt = Build.A.Receipt.WithBlockNumber(blockNumber).WithTransactionHash(TestItem.KeccakA).WithIndex(0).TestObject;
@@ -407,7 +407,7 @@ namespace Nethermind.JsonRpc.Test.Modules.Subscribe
         [Test]
         public void TransactionReceiptsSubscription_dispose_stops_delivery()
         {
-            int blockNumber = 55555;
+            ulong blockNumber = 55555;
             BlockHeader blockHeader = Build.A.BlockHeader.WithNumber(blockNumber).TestObject;
 
             TransactionReceiptsSubscription subscription = new(
