+++ src/Nethermind/Nethermind.State.Flat.Test/Sync/Snap/FlatStateRootIndexTests.cs
@@ -30,7 +30,7 @@ public class FlatStateRootIndexTests
         return blocks;
     }
 
-    private static (BlockTree tree, List<Block> blocks, FlatStateRootIndex index) BuildIndex(int chainLength, int lastN = 10)
+    private static (BlockTree tree, List<Block> blocks, FlatStateRootIndex index) BuildIndex(int chainLength, ulong lastN = 10)
     {
         List<Block> blocks = BuildChain(chainLength);
         BlockTree tree = Build.A.BlockTree().WithBlocks(blocks.ToArray()).TestObject;
